Implemented Polyline Path geofencing
Now there is a colored border on the sides of the map.
Red Color indicates that user is not following the path.
Green Color indicates correct following of the path by user.
Whenever user steps out of the polyline path, a dialog is shown alerting the user of his actions and actions are suggeseted.